test(atlas): e2e mixed quantization test on TinyLlama
Adds TestMixedModeE2E, a real (not mocked) end-to-end test that runs
the full mixed-precision pipeline against TinyLlama-1.1B on real
hardware: profile, plan, mixed-quantize, eval, pack.

Also fixes a real precision bug surfaced by running the existing
uniform-mode e2e test on hardware: PerplexityEval._compute_perplexity
summed cross_entropy loss in bfloat16, which silently rounds to
integer-like values once the per-sample summed loss exceeds ~64
(bfloat16 has ~3 significant decimal digits). Casting logits to
float32 before the loss sum fixes this; uniform 4-bit PPL delta on
TinyLlama is now a sane 4.3% instead of an inflated, deterministic
105%.

Also fixes TestEndToEnd, the pre-existing uniform-mode e2e test: it
called pipeline.run(...) without mode=, and Task 4 changed
Pipeline.run's default to mode="mixed", so the test was silently
exercising the mixed-quantization path instead of uniform. Pinned it
to mode="uniform" to restore its original intent.

TestMixedModeE2E itself currently fails (ppl_delta_pct ~105% instead
of <50%): real-hardware testing surfaced a genuine bug in Task 1's
LayerProfiler — its sensitivity score is a min-max normalization of
raw residual-stream activation norms, which grow near-monotonically
with depth in transformer architectures (an architectural artifact,
not a quantization-sensitivity signal). This causes QuantPlanner to
assign 2-bit quantization to the earliest, most quantization-critical
layers. Flagged as a follow-up task; not fixed here since it requires
redesigning Task 1's sensitivity metric and its locked-in unit test
contract.

+++ atlas/eval/perplexity.py
@@ -71,7 +71,7 @@ def _compute_perplexity(model_path: str, samples: list[str]) -> float:
         targets = mx.array(tokens[1:])
 
         logits = model(input_ids)
-        logits = logits.squeeze(0)
+        logits = logits.squeeze(0).astype(mx.float32)
 
         loss = nn.losses.cross_entropy(logits, targets, reduction="sum")
         total_loss += loss.item()
